+++ src/index.ts
@@ -1,4 +1,12 @@
 import CachedHandler from './CachedHandler';
 export default CachedHandler;
+
 import RedisStringsHandler from './RedisStringsHandler';
 export { RedisStringsHandler };
+export type { CreateRedisStringsHandlerOptions } from './RedisStringsHandler';
+
+import {
+  redisCacheHandler,
+  getRedisCacheComponentsHandler,
+} from './CacheComponentsHandler';
+export { redisCacheHandler, getRedisCacheComponentsHandler };